Add Artist component.
The artist will be the component that uses the Pen and Grid components
to carry out commands sent to it. At the moment, i’en just implemented
the ‘line’ command. This framework can later be extended to include
other commands (e.g., arc, circle, dashed-line, etc).

The artist, like the grid, assumes that the plotter starts out at the
origin/zeroed at the top left corner of the total plottable surface.